New storage system (part 1)

+++ src/actions/files.js
@@ -11,14 +11,8 @@ export const checkFiles = () => {
   for (var key in paths.files) {
     let file = paths.files[key]
 
-    if (typeof file === 'object') {
-      if (!fs.existsSync(file.path)) {
-        fs.writeFileSync(file.path, JSON.stringify(file.defaultContent))
-      }
-    } else {
-      if (!fs.existsSync(file)) {
-        fs.writeFileSync(file, '[]')
-      }
+    if (typeof file === 'object' && file.autoCreate && !fs.existsSync(file.path)) {
+      fs.writeFileSync(file.path, JSON.stringify(file.defaultContent))
     }
   }
 }
